Added LocallyCastedActor
LocallyCastedActor Class is used for every object that can be casted locally on the map using luma. Cast ability for local casts isn't realized yet. Widgets are updated to display local cast at the top of the screen. Using overlay material + opacity mask on object to show it's presence without actually destroying it.