Map updated
Building “stage with container” moved to the correct location according to the street, passage to the building changed. Added underground tunnel. Added borders of the destroyed corporat base, moved some containers. Changed the path along the interior of the building in front of the corporat tower, added a path behind the loot with climbing. Scaffolding is no longer active on the statue street, for scenery.